% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSupport.R
\name{CI_draw}
\alias{CI_draw}
\title{A function to draw the confidence interval as a diamond}
\usage{
CI_draw(plotobj, x, y, a = 1, b = 1)
}
\arguments{
\item{plotobj}{the obj to be plot}

\item{x}{the x coordinate of the center to be plotted}

\item{y}{the y coordinate of the center to be plotted}

\item{a}{the distance between the center to the vertext on x-axis of the diamond}

\item{b}{the distance between the center to the vertext on x-axis of the diamond}
}
\value{
a ggplot object
}
\description{
A function to draw the confidence interval as a diamond
}
\keyword{internal}
