% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_tree_Support.R
\name{compute_rebetQ}
\alias{compute_rebetQ}
\title{A function to compute RE Q-between}
\usage{
compute_rebetQ(y, vi, xk)
}
\arguments{
\item{y}{effect size}

\item{vi}{sampling variance}

\item{xk}{moderator}
}
\value{
Q-between and tau2
}
\description{
A function to compute RE Q-between
}
\keyword{internal}
