% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat.balanced}
\alias{dat.balanced}
\title{A simulated meta-analytic data set with balanced pure interaction effects}
\format{A data frame of 60 studies with 4 moderators
\itemize{
  \item efk: The effect size of each study expressed as Hedges' g
  \item vark: The sampling variance of the effect size
  \item x1 to x4: Four randomly generated moderators. x1, x2, and x4 are dichotomous variables,
  x3 is a continuous variable generated from uniform distribution.
  }}
\usage{
data(dat.balanced)
}
\description{
Data simulated from a true model with pure interactions between two
moderators: x1, x2. If x1 = 0 and x2 = 1 or x1 = 1 and x2 = 0, 
the true effect size is 0.50. Otherwise, the true effect size is 0.
}
\keyword{datasets}
\keyword{simulated}
