% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.compute_re_Q_}
\alias{.compute_re_Q_}
\title{Compute re Q for different values of tau2}
\usage{
.compute_re_Q_(x1, x2, x3, x4, xuni, x5, x6)
}
\arguments{
\item{x1}{the effect size g in the unsplit leaves}

\item{x2}{the sampling variance vi in the unsplit leaves}

\item{x3}{the labels of nodes in the unsplit leaves}

\item{x4}{tau2}

\item{xuni}{the unique labels in the unsplit leaves}

\item{x5}{the effect size g in the parent leaf}

\item{x6}{the sampling variance vi in the parent leaf}
}
\description{
Compute re Q for different values of tau2
}
\keyword{internal}
