% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.compute_tau_}
\alias{.compute_tau_}
\title{Compute tau2}
\usage{
.compute_tau_(x1, x2, x3, xuni, x4, x5)
}
\arguments{
\item{x1}{the effect size g in the unsplit leaves}

\item{x2}{the sampling variance vi in the unsplit leaves}

\item{x3}{the lable of the unsplit leaves}

\item{xuni}{the unique labels in the unsplit leaves}

\item{x4}{the sorted effect size in the parent leaf}

\item{x5}{the sorted sampling variance in the parent leaf}
}
\description{
Compute tau2
}
\keyword{internal}
