% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ComputeY}
\alias{.ComputeY}
\title{Compute the subgroup effect sizes}
\usage{
.ComputeY(x1, y, vi, tau2)
}
\arguments{
\item{x1}{the node labels for each study}

\item{y}{the effect size}

\item{vi}{the sampling variance}

\item{tau2}{the residual heterogeneity}
}
\description{
Compute the subgroup effect sizes
}
\keyword{internal}
