% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.partition}
\alias{.partition}
\title{Partition the test set based on a trained tree}
\usage{
.partition(x1, x2, x3, x4, x5, x6)
}
\arguments{
\item{x1}{the tree component of the REmrt object}

\item{x2}{the moderators in the test set}

\item{x3}{indicates whether a moderator is numeric or not}

\item{x4}{the index vector of the spliting moderators}

\item{x5}{the list of split points}

\item{x6}{the moderators in the training set}
}
\description{
Partition the test set based on a trained tree
}
\keyword{internal}
