% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.PredY}
\alias{.PredY}
\title{Predict effect size for the test set}
\usage{
.PredY(x1, x2)
}
\arguments{
\item{x1}{the list of subgroup means}

\item{x2}{predicted subgroup membership for the test set}
}
\description{
Predict effect size for the test set
}
\keyword{internal}
