% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.ReplaceNA}
\alias{.ReplaceNA}
\title{Replace missing values by the overall weighted mean}
\usage{
.ReplaceNA(x1, x2, y, vi, tau2)
}
\arguments{
\item{x1}{the two-column matrix of the indices of missing values}

\item{x2}{the matrix of predicted y with missing values}

\item{y}{the effect size}

\item{vi}{the sampling variance}

\item{tau2}{the residual heterogeneity}
}
\description{
Replace missing values by the overall weighted mean
}
\keyword{internal}
