% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSupport.R
\name{encodeHtml}
\alias{encodeHtml}
\title{A function to deal with symbols}
\usage{
encodeHtml(input)
}
\arguments{
\item{input}{a string}
}
\value{
converted string
}
\description{
A function to deal with symbols
}
\keyword{internal}
