% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_tree_Support.R
\name{find_second_split}
\alias{find_second_split}
\title{A function to find the optimal combination of first two splits, 
and the corresponding Q-between given the first split}
\usage{
find_second_split(xk, first.splits, y, vi, minbucket, minsplit)
}
\arguments{
\item{xk}{moderator vector}

\item{first.splits}{possible first splits}

\item{y}{effect size}

\item{vi}{sampling variance}

\item{minbucket}{the minimal number of studies in child nodes}

\item{minsplit}{the minimal number of studies in parent node}
}
\value{
a list including all possible combinations of the triplet
}
\description{
A function to find the optimal combination of first two splits, 
and the corresponding Q-between given the first split
}
\keyword{internal}
