% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_tree_Support.R
\name{find_triplet}
\alias{find_triplet}
\title{A function to find the best triplets of parent, moderator, and split point.}
\usage{
find_triplet(xk, nodeMbrship, y, vi, minbucket, minsplit)
}
\arguments{
\item{xk}{moderator vector}

\item{nodeMbrship}{node membership vector}

\item{y}{effect size}

\item{vi}{sampling variance}

\item{minbucket}{the minimal number of studies in child nodes}

\item{minsplit}{the minimal number of studies in parent node}
}
\value{
pleaf: the parent node

cstar: the split point

rank: the rank used to order categorical moderator
}
\description{
A function to find the best triplets of parent, moderator, and split point.
}
\keyword{internal}
