% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow_tree_Support.R
\name{make_first_split}
\alias{make_first_split}
\title{A function to list all possible split points for the first split}
\usage{
make_first_split(xk, y)
}
\arguments{
\item{xk}{moderator}

\item{y}{effect size}
}
\value{
childNodes: child nodes membership

rank: the rank used to order categorical moderator

is.num: if the moderator is numeric
}
\description{
A function to list all possible split points for the first split
}
\keyword{internal}
