% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSupport.R
\name{oval_draw}
\alias{oval_draw}
\title{A function to draw an oval}
\usage{
oval_draw(plotobj, x, y, c, x.scale = 1, y.scale = 1, ...)
}
\arguments{
\item{plotobj}{the obj to be plot}

\item{x}{x}

\item{y}{y}

\item{c}{c}

\item{x.scale}{x.scale}

\item{y.scale}{y.scale}
}
\value{
a ggplot object
}
\description{
A function to draw an oval
}
\keyword{internal}
