% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.FEmrt}
\alias{plot.FEmrt}
\title{Visualisation of a FE meta-tree}
\usage{
\method{plot}{FEmrt}(x, ...)
}
\arguments{
\item{x}{A FEmrt object.}

\item{...}{Arguments that pass to prp()}
}
\description{
Plot function for a \code{FEmrt} object. The plot shows the result of \code{FEmrt}.
The plot function uses the plot method from the package \pkg{rpart.plot} of Stephen Milborrow
(2016).
}
\details{
For categorical variables we recommend to use short names for levels to avoid overlapping
labels at split points.
For each split of the tree it holds that if the condition (in the parent node) is satisfied,
then the study goes to the left node and if not it goes to the right node.
}
