% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRE.R
\name{plot.REmrt}
\alias{plot.REmrt}
\title{Visualisation of a RE meta-tree}
\usage{
\method{plot}{REmrt}(x, ...)
}
\arguments{
\item{x}{A REmrt object.}

\item{...}{Additional arguments to pass.}
}
\description{
Plot function for a \code{REmrt} object. The plot shows the result of \code{REmrt}.
The plot function uses the plot method from the package \pkg{ggplot2}
}
\details{
For categorical variables we recommend to use short names for levels to avoid overlapping labels at split points.
}
