% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.FEmrt}
\alias{predict.FEmrt}
\title{Predictions from a fitted metacart object}
\usage{
\method{predict}{FEmrt}(object, newdata, ...)
}
\arguments{
\item{object}{fitted model object of class "FEmrt".}

\item{newdata}{data frame containing the values at which predictions are required.}

\item{...}{Arguments that pass to other methods.}
}
\value{
A data frame containing the predicted effect size, the moderators, and the corresponding node lables in the fitted tree.
}
\description{
Returns a data frame of predicted effect sizes and moderators from a fitted metacart object
}
