% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.FEmrt}
\alias{print.FEmrt}
\title{Print function for FEmrt}
\usage{
\method{print}{FEmrt}(x, ...)
}
\arguments{
\item{x}{fitted tree of class \code{FEmrt}.}

\item{\dots}{additional arguments to be passed.}
}
\description{
Print the results of a FEmrt object
}
\details{
The function returns the objects concerning the analysis results.
}
