% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe_permute.R
\name{Q_selected_size_GS}
\alias{Q_selected_size_GS}
\title{A function returns the Q-between from the tree with given size}
\usage{
Q_selected_size_GS(tree, nsplit, mods, minbucket, minsplit, ...)
}
\arguments{
\item{tree}{A initial tree fitted by rpart, needs to an rpart object.}

\item{nsplit}{the required number of splits}

\item{mods}{the moderators found by the tree.}

\item{minbucket}{the minimum number of observations in any terminal <leaf> node.}

\item{minsplit}{the minimum number of observations that must exist in a node in order for a split to be attempted.}

\item{...}{Additional arguments passed to prune.rpart().}
}
\value{
The pruned tree
}
\description{
A function returns the Q-between from the tree with given size
}
\keyword{internal}
