% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REtreeSupport.R
\name{rebetQ}
\alias{rebetQ}
\title{A function to compute Q-between and residual heterogeneity}
\usage{
rebetQ(yi, vi, mods)
}
\arguments{
\item{yi}{the effect sizes}

\item{vi}{the sampling variances}

\item{mods}{the subgrouping moderator}
}
\value{
Q-between and the residual heterogeneity under mixed effects model assumption
}
\description{
A function to compute Q-between and residual heterogeneity
}
\keyword{internal}
