% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REtreeSupport.R
\name{REmrt.fit0}
\alias{REmrt.fit0}
\title{A function to fit the tree for cross-validation}
\usage{
REmrt.fit0(mf, maxL)
}
\arguments{
\item{mf}{the data.frame to grow the tree}

\item{maxL}{the maximum number of splits}
}
\value{
a list including a tree, the split points, the data, and the nodes after each split
}
\description{
A function to fit the tree for cross-validation
}
\keyword{internal}
