% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REtreeSupport.R
\name{REmrt.fit1}
\alias{REmrt.fit1}
\title{A function to grow the tree}
\usage{
REmrt.fit1(mf, maxL, minsplit = 2, delQ = 0.001)
}
\arguments{
\item{mf}{the data.frame to grow the tree}

\item{maxL}{the maximum number of splits}

\item{delQ}{the stopping rule for decrease of between-subgroups Q. Any split that does not decrease the between-subgroups Q is not attempted.}

\item{minisplit}{the minimal number of studies in a parent node to be split}
}
\value{
a list including a tree, the split points, the data, and the nodes after each split
}
\description{
A function to grow the tree
}
\keyword{internal}
