% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REtreeSupport.R
\name{REmrt.prednode}
\alias{REmrt.prednode}
\title{A function to subgroup newdata according to the fitted tree}
\usage{
REmrt.prednode(x, newdata)
}
\arguments{
\item{x}{the fitted RE meta-tree results}

\item{newdata}{the new data set}
}
\value{
a matrix consists of node lables
}
\description{
A function to subgroup newdata according to the fitted tree
}
\keyword{internal}
