% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REtreeSupport.R
\name{REmrt.xvalid}
\alias{REmrt.xvalid}
\title{A function to perform cross-validation for RE meta-tree}
\usage{
REmrt.xvalid(mf, maxL, n.fold = 10)
}
\arguments{
\item{mf}{the data.frame to grow the tree}

\item{maxL}{the maximum number of splits}

\item{n.fold}{the number of folds to perform the cross-validation}
}
\value{
a cp table
}
\description{
A function to perform cross-validation for RE meta-tree
}
\keyword{internal}
