% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REmrt_look_ahead.R
\name{REmrt_GS_}
\alias{REmrt_GS_}
\title{A function to fit the tree with look-ahead option}
\usage{
REmrt_GS_(mf, maxL, minbucket, minsplit, cp, lookahead)
}
\arguments{
\item{mf}{the data.frame to grow the tree}

\item{maxL}{the maximum number of splits}

\item{minbucket}{the minimum number of the studies in a terminal node}

\item{minsplit}{the minimal number of studies in a parent node to be split}

\item{cp}{the stopping rule for decrease of between-subgroups Q. Any split that does not decrease the between-subgroups Q is not attempted.}

\item{lookahead}{an argument indicating whether to apply the "look-ahead" strategy when fitting the tree}
}
\value{
a list including a tree, the split points, the data, and the nodes after each split
}
\description{
A function to fit the tree with look-ahead option
}
\keyword{internal}
