% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.FEmrt}
\alias{summary.FEmrt}
\title{Summary of the results of a FE meta-tree object}
\usage{
\method{summary}{FEmrt}(object, digits = 3, ...)
}
\arguments{
\item{object}{fitted tree of class \code{FEmrt}.}

\item{digits}{specified number of decimals in the printed results.}

\item{\dots}{additional arguments to be passed.}
}
\description{
Summary of the results of a FE meta-tree object
}
\details{
If no moderator effect is detected,
the summary function will show the standard meta-analysis results.
Otherwise, the summary function will show the subgroup meta-analysis results,
with the significance test results for moderator effects, the split points of the moderators,
and the estimated subgroup summary effect sizes.
}
