% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_tree_Support.R
\name{treepruner}
\alias{treepruner}
\title{Prune a tree}
\usage{
treepruner(tree, c, ...)
}
\arguments{
\item{tree}{A initial tree fitted by rpart, needs to an rpart object.}

\item{c}{A scalar to prune the  tree by selecting the tree with minum cross-validation error plus the standard error multiplied by c.}

\item{...}{Additional arguments passed to prune.rpart().}
}
\value{
The pruned tree
}
\description{
Prune an initial rpart tree by "c-standard-error" rule.
}
\keyword{internal}
