% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSupport.R
\name{updateNodes}
\alias{updateNodes}
\title{A function to update node}
\usage{
updateNodes(nodes, newNode, name = "leaf")
}
\arguments{
\item{nodes}{a data frame}

\item{newNode}{a new row}

\item{name}{a string}
}
\value{
a data.frame updated the nodes
}
\description{
A function to update node
}
\keyword{internal}
