% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_valid_Support.R
\name{Xvalid_all}
\alias{Xvalid_all}
\title{A function to compute cross-validation errors}
\usage{
Xvalid_all(func, mf, maxL, n.fold, minbucket, minsplit, cp, lookahead)
}
\arguments{
\item{func}{the function used to fit the model}

\item{mf}{the data set with formula specified}

\item{maxL}{the maximum number of splits}

\item{n.fold}{the number of folds}

\item{minbucket}{the minimum number of the studies in a terminal node}

\item{minsplit}{the minimal number of studies in a parent node to be split}

\item{cp}{the stopping rule for decrease of between-subgroups Q. Any split that does not decrease the between-subgroups Q is not attempted.}

\item{lookahead}{an argument indicating whether to apply the "look-ahead" strategy when fitting the tree}
}
\description{
A function to compute cross-validation errors
}
\keyword{internal}
