% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_GS_lookahead_Support.R
\name{make_first_split}
\alias{make_first_split}
\title{A function to list all possible split points for the first split}
\usage{
make_first_split(xk, y, minbucket)
}
\arguments{
\item{xk}{moderator}

\item{y}{effect size}

\item{minbucket}{the minimum number of the studies in a terminal node}
}
\value{
childNodes child nodes membership

rank the rank used to order a categorical moderator

is.num indicates if the moderator is numeric
}
\description{
A function to list all possible split points for the first split
}
\keyword{internal}
