% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFE.R
\name{plot.FEmrt}
\alias{plot.FEmrt}
\title{Visualisation of a FE meta-tree}
\usage{
\method{plot}{FEmrt}(x, ...)
}
\arguments{
\item{x}{A FEmrt object.}

\item{...}{additional arguments to pass}
}
\description{
Plot function for a \code{FEmrt} object. The plot shows the result of \code{FEmrt}.
The plot function uses the plot method from the package \pkg{ggplot2}
}
\details{
For categorical variables we recommend to use short names for levels to avoid overlapping
labels at split points.
}
