% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_valid_Support.R
\name{prednode_cpp}
\alias{prednode_cpp}
\title{A function to parition newdata based on the fitted model}
\usage{
prednode_cpp(x, newdata)
}
\arguments{
\item{x}{a fitted metaCART model}

\item{newdata}{new data for prediction}
}
\description{
A function to parition newdata based on the fitted model
}
\keyword{internal}
