% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.REmrt}
\alias{print.REmrt}
\title{Print function for REmrt}
\usage{
\method{print}{REmrt}(x, ...)
}
\arguments{
\item{x}{fitted tree of class \code{FEmrt}.}

\item{\dots}{additional arguments to be passed.}
}
\description{
Print the results of a REmrt object
}
\details{
The function returns the results (e.g., the value of the Q-between) after each split of the tree.
}
\examples{
data(SimData)
test <- REmrt(efk~m1+m2+m3+m4+m5, vark, data = SimData, maxL = 2, c= 0.5)
print(test)
}
